requirements.txt
# Core
crewai>=0.70.0
langchain>=0.1.0
langchain-community>=0.1.0

# LLM
litellm>=1.35.0

# Tools
requests>=2.31.0
beautifulsoup4>=4.12.0
bibtexparser>=1.4.0
doi2bib3>=0.4.0

# Configuration
python-dotenv>=1.0.0